requirements.txt
3-1==1.0.0
altgraph==0.17
asgiref==3.5.2
autopep8==1.5.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bokeh==2.1.1
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.4
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.0.4
django-bootstrap-static==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==1.9.2
django-filter==21.1
django-heroku==0.3.1
django-restframework==0.0.1
django-secure==1.0.2
django-social-share==2.3.0
djangorestframework==3.13.1
ezdxf==0.17.1b3
fbs==0.8.9
filelock==3.8.0
fonttools==4.33.3
future==0.18.2
gTTS==2.2.4
gunicorn==20.0.4
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
Jinja2==2.11.2
kiwisolver==1.4.2
macholib==1.14
Markdown==3.3.7
MarkupSafe==1.1.1
matplotlib==3.5.2
moviepy==1.0.3
mysqlclient==2.0.1
numpy==1.22.4
oauthlib==3.2.0
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.1.1
pefile==2019.4.18
Pillow==9.1.1
platformdirs==2.5.2
proglog==0.1.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.21
PyInstaller==3.4
pyinstaller-hooks-contrib==2020.7
PyJWT==2.4.0
PyMySQL==0.10.0
pyparsing==3.0.9
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.5.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.1
tornado==6.0.4
tqdm==4.64.0
typing-extensions==3.7.4.2
urllib3==1.26.9
virtualenv==20.16.5
whitenoise==5.2.0
zipp==3.8.0S
qrcode==7.3.1